My life is falling apart because I can't Qt and I was stubborn.
Realizing that (learning a new library) && (trying to complete a project
using new materials and concepts) = false;
||=== Build failed: 1 error(s), 0 warning(s) (0 minute(s), 0 second(s))
===|

But there is hope in C++.  I have learned the error of my ways and
return to my first love,  if C will but have me. That was a play on
words. The end.