Addition of request
Things are beginning to come together a bit more. I am ATTEMPTING to
implement a command design pattern. We shall see hoe it goes. The
request class can now process user selection and is capable of offering
minimal input verification. The expansion of these abilities may come
later but there is much that needs to be integrated into the system.
Lots of programming to be done.