Invoker Progression
I have gone through the requirements documents and made some much needed
notation on the status of the simulator. As of right now I am going to
create an additional project for the pcb test suite. The test suite will
contain all of the requirements and commands Once that is fully
functioning the entire suite will be added to the project and be run
from the runProgram file. Godspeed.