Progress isn't my middle name, Rose is.
We have begun the connection of the request to the invoker class. A
goodbye message has been added to the exit task. It's cute if I may say
so myself. AND I MAY!
The next step: initiate queue construction with the selection of the
invoker
all of the necessary options should already exist since the pcb was
constructed by itself outside of the simulator and tested. But I'm not
one for avoiding crisis.

After that:  PART 3, YIPEEE. we have a solid plan (why do I keep saying
we? this really is a solo project. Maybe I mean Han (if it's a solo
project :P ! Ha ha. I joke with myself when pushing to github. This
probably means I'm a loser)